Move GPS files from edge firmware to libraries
