+++ lib/gps-tracker/EdgePlatform.cpp
