+++ lib/gps-tracker/EdgePlatform.h
