+++ lib/gps-tracker/location_service.cpp
