+++ lib/gps-tracker/location_service.h
