+++ lib/gps-tracker/tracker_config.h
